Update setup instructions using new `KitConfig`
With the release of Proximity Kit for Android 0.10.0 the way the manager
is setup has changed. This shows how to load the `.properties` files and
then supply the resulting configuration to the `ProximityKitManager`.
This change marks the first step toward Proximity Kit for Android
allowing multiple manager instances (we do not support this yet, but
this API change reflects the work towards that goal).

Additionally, we should clients how to properly create singleton
instances so that they can ensure they have thread safe creation of
their manager instance.